augment wiki data
When wiki json file is loaded if it hasn't been augmented yet, it loads the web page and augments the text to include information about the associated HTML tags for that text.  It then resaves the file.   It takes a little while to run but only needs to be run once.

Not doing anything with augmented data yet but will be employed to create better section consolidation for grounding